+++ resources/lang/cs_CZ/validation.php
@@ -67,6 +67,8 @@ return [
     'not_transfer_account'           => 'Tento účet není účet, který lze použít pro převody.',
     'require_currency_amount'        => 'Obsah tohoto pole je neplatný bez informace o měně.',
     'require_foreign_currency'       => 'This field requires a number',
+    'require_foreign_dest'           => 'This field value must match the currency of the destination account.',
+    'require_foreign_src'            => 'This field value must match the currency of the source account.',
     'equal_description'              => 'Popis transakce nesmí být stejný jako globální popis.',
     'file_invalid_mime'              => 'Soubor ":name" je typu ":mime", který není schválen pro nahrání.',
     'file_too_large'                 => 'Soubor ":name" je příliš velký.',