+++ resources/lang/sk_SK/validation.php
@@ -67,6 +67,8 @@ return [
     'not_transfer_account'           => 'Tento účet nie je účet, ktorý je možné použiť pre prevody.',
     'require_currency_amount'        => 'Obsah tohto poľa je bez informácie o cudzej mene neplatný.',
     'require_foreign_currency'       => 'This field requires a number',
+    'require_foreign_dest'           => 'This field value must match the currency of the destination account.',
+    'require_foreign_src'            => 'This field value must match the currency of the source account.',
     'equal_description'              => 'Popis transakcie nesmie byť rovnaký ako globálny popis.',
     'file_invalid_mime'              => 'Súbor ":name" je typu ":mime", ktorý nie je pre nahrávanie schválený.',
     'file_too_large'                 => 'Súbor ":name" je príliš veľký.',