+++ resources/lang/sl_SI/validation.php
@@ -67,6 +67,8 @@ return [
     'not_transfer_account'           => 'This account is not an account that can be used for transfers.',
     'require_currency_amount'        => 'Vsebina tega polja ni veljavna brez podatkov o tujih zneskih.',
     'require_foreign_currency'       => 'This field requires a number',
+    'require_foreign_dest'           => 'This field value must match the currency of the destination account.',
+    'require_foreign_src'            => 'This field value must match the currency of the source account.',
     'equal_description'              => 'Transaction description should not equal global description.',
     'file_invalid_mime'              => 'File ":name" is of type ":mime" which is not accepted as a new upload.',
     'file_too_large'                 => 'File ":name" is too large.',